Coin, CoinSpawner, TextOverlay
Coins: when the player touches them, they gain 1 coin, and the coin is destroyed.
Coin spawners: spawn coins based on a random number parameter.  For instance, if you give the parameter "3", it will have a one in three chance of spawning a coin.  After spawning the coin, the spawner is destroyed.
TextOverlay: instantiated by the player, shows their coin count over their head.